requirements.txt
# Add your project dependencies here
selenium
pillow
schedule